Restructure and namespace to allow for composer autoloading
