+++ composer.json
@@ -32,5 +32,10 @@
 	},
 	"require-dev": {
 	    "phpunit/phpunit": "4.5.*"
+	},
+	"autoload": {
+		"psr-4": {
+			"Mike42\\": "src/Mike42"
+		}
 	}
 }
