+++ src/Mike42/Escpos/CapabilityProfiles/EposTepCapabilityProfile.php
@@ -1,4 +1,6 @@
 <?php
+namespace Mike42\Escpos\CapabilityProfiles;
+
 class EposTepCapabilityProfile extends DefaultCapabilityProfile {
 	// TODO override list of code pages
 }
