+++ src/Mike42/Escpos/PrintBuffers/cache/Characters-StarCapabilityProfile.ser.gz
