+++ src/Mike42/Escpos/PrintConnectors/PrintConnector.php
@@ -1,4 +1,5 @@
 <?php
+namespace Mike42\Escpos\PrintConnectors;
 /**
  * escpos-php, a Thermal receipt printer library, for use with
  * ESC/POS compatible printers.
@@ -27,46 +28,24 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  * 
- * PrintConnector for passing print data to a file.
+ * Interface passed to Escpos class for receiving print data. Print connectors
+ * are responsible for transporting this to the actual printer.
  */
-class FilePrintConnector implements PrintConnector {
+interface PrintConnector {
 	/**
-	 * @var resource The file pointer to send data to.
+	 * Print connectors should cause a NOTICE if they are deconstructed
+	 * when they have not been finalized.
 	 */
-	protected $fp;
+	public function __destruct();
 
 	/**
-	 * Construct new connector, given a filename
-	 * 
-	 * @param string $filename
+	 * Finish using this print connector (close file, socket, send
+	 * accumulated output, etc).
 	 */
-	public function __construct($filename) {
-		$this -> fp = fopen($filename, "wb+");
-		if($this -> fp === false) {
-			throw new Exception("Cannot initialise FilePrintConnector.");
-		}
-	}
-	
-	public function __destruct() {
-		if($this -> fp !== false) {
-			trigger_error("Print connector was not finalized. Did you forget to close the printer?", E_USER_NOTICE);
-		}
-	}
+	public function finalize();
 
 	/**
-	 * Close file pointer
-	 */
-	public function finalize() {
-		fclose($this -> fp);
-		$this -> fp = false;
-	}
-	
-	/**
-	 * Write data to the file
-	 * 
 	 * @param string $data
 	 */
-	public function write($data) {
-		fwrite($this -> fp, $data);
-	}
+	public function write($data);
 }
