+++ src/Mike42/Escpos/PrintConnectors/WindowsPrintConnector.php
@@ -1,4 +1,5 @@
 <?php
+namespace Mike42\Escpos\PrintConnectors;
 /**
  * escpos-php, a Thermal receipt printer library, for use with
  * ESC/POS compatible printers.
